Added new  functionality :
Now it is possible to create a simulation that, after the time ends, processes only events that are not discardable.

Functionalities have been embedded in new APIs :

create_simulation_until_end : creates the new kind of simulation

createUndiscardableEvent : creates an undiscardable event.

So default creations' behaviors remain the usual ones

Co-authored-by: Alessandro Lioi <lioia@users.noreply.github.com>